error stack added

+++ src/app/modules/product/bicycle.controller.ts
@@ -1,13 +1,17 @@
-import { Request, Response } from "express";
+import { NextFunction, Request, Response } from "express";
 import { bicycleZodSchema } from "./bicycle.interface";
 import BicycleModel from "./bicycle.model";
-import { BicycleServices } from "./bicycle.service"; // Import the BicycleServices
+import { BicycleServices } from "./bicycle.service";
 
-const createBicycle = async (req: Request, res: Response) => {
+const createBicycle = async (
+	req: Request,
+	res: Response,
+	next: NextFunction,
+) => {
 	try {
 		const bicycleData = req.body;
 
-		// validate bicycle data before sending to database
+		// Validate bicycle data before sending to the database
 		const validatedData = bicycleZodSchema.parse(bicycleData);
 
 		// Call to BicycleServices to create a bicycle
@@ -19,34 +23,45 @@ const createBicycle = async (req: Request, res: Response) => {
 			data: result,
 		});
 	} catch (error) {
-		res.status(400).json({
-			success: false,
-			message: "Something went wrong with bicycle creation.",
-			error,
-		});
+		next(error);
 	}
 };
 
-const getAllBicycles = async (req: Request, res: Response) => {
+const getAllBicycles = async (
+	req: Request,
+	res: Response,
+	next: NextFunction,
+) => {
 	try {
-		const result = await BicycleServices.getAllBicyclesFromDB();
+		const { name, brand, type } = req.query;
+		const query: { name?: object; brand?: object; type?: object } = {};
 
+		if (name) {
+			query.name = { $regex: name, $options: "i" };
+		}
+		if (brand) {
+			query.brand = { $regex: brand, $options: "i" };
+		}
+		if (type) {
+			query.type = { $regex: type, $options: "i" };
+		}
+
+		const result = await BicycleServices.getAllBicyclesFromDB(query);
 		res.status(200).json({
 			success: true,
-			message: "Bicycles are retrieved successfully",
+			message: "Bicycles retrieved successfully",
 			data: result,
 		});
 	} catch (error) {
-		console.log(error);
-		res.status(500).json({
-			success: false,
-			message: "Something went wrong while retrieving bicycles.",
-			error,
-		});
+		next(error);
 	}
 };
 
-const getSingleBicycle = async (req: Request, res: Response) => {
+const getSingleBicycle = async (
+	req: Request,
+	res: Response,
+	next: NextFunction,
+) => {
 	try {
 		const { productId } = req.params;
 
@@ -58,21 +73,19 @@ const getSingleBicycle = async (req: Request, res: Response) => {
 
 		res.status(200).json({
 			success: true,
-			message: "Bicycle is retrieved successfully",
+			message: "Bicycle retrieved successfully",
 			data: result,
 		});
-		// eslint-disable-next-line @typescript-eslint/no-explicit-any
-	} catch (error: any) {
-		res.status(500).json({
-			success: false,
-			message:
-				error.message || "Something went wrong while retrieving the bicycle.",
-			error,
-		});
+	} catch (error) {
+		next(error);
 	}
 };
 
-const updateBicycle = async (req: Request, res: Response) => {
+const updateBicycle = async (
+	req: Request,
+	res: Response,
+	next: NextFunction,
+) => {
 	try {
 		const { productId } = req.params;
 		const bicycleData = req.body;
@@ -81,10 +94,10 @@ const updateBicycle = async (req: Request, res: Response) => {
 			throw new Error("This bicycle does not exist.");
 		}
 
-		// validate bicycle data before sending to database
+		// Validate bicycle data before sending to the database
 		const validatedData = bicycleZodSchema.partial().parse(bicycleData);
 
-		// Call to BicycleServices to create a bicycle
+		// Call to BicycleServices to update a bicycle
 		const result = await BicycleServices.updateBicycleFromDB(
 			productId,
 			validatedData,
@@ -92,36 +105,37 @@ const updateBicycle = async (req: Request, res: Response) => {
 
 		res.status(200).json({
 			success: true,
-			message: "Bicycle is updated successfully",
-			data: result,
-		});
-		// eslint-disable-next-line @typescript-eslint/no-explicit-any
-	} catch (error: any) {
-		res.status(400).json({
-			success: false,
-			message: error.message || "Something went wrong during bicycle update.",
-			error,
-		});
-	}
-};
-
-const deleteBicycle = async (req: Request, res: Response) => {
-	try {
-		const { productId } = req.params;
-
-		const result = await BicycleServices.deleteBicycleFromDB(productId);
-
-		res.status(200).json({
-			success: true,
-			message: "Bicycle is deleted successfully",
+			message: "Bicycle updated successfully",
 			data: result,
 		});
 	} catch (error) {
-		res.status(400).json({
-			success: false,
-			message: "Something went wrong during bicycle delection.",
-			error,
-		});
+		next(error);
+	}
+};
+
+const deleteBicycle = async (
+	req: Request,
+	res: Response,
+	next: NextFunction,
+) => {
+	try {
+		const { productId } = req.params;
+
+		if (!(await BicycleModel.isExist(productId))) {
+			throw new Error("This bicycle does not exist.");
+		}
+
+		const result = await BicycleServices.deleteBicycleFromDB(productId);
+
+		if (result.modifiedCount === 1) {
+			res.status(200).json({
+				success: true,
+				message: "Bicycle deleted successfully",
+				data: {},
+			});
+		}
+	} catch (error) {
+		next(error);
 	}
 };
